+++ libs/qdarktheme/__init__.py
@@ -1,0 +1,6 @@
+"""PyQtDarkTheme - A flat dark theme for PySide and PyQt.
+
+- Repository: https://github.com/5yutan5/PyQtDarkTheme
+- Documentation: https://pyqtdarktheme.readthedocs.io
+"""
+from libs.qdarktheme.main import __version__, clear_cache, get_themes, load_palette, load_stylesheet